Refactor: extract app from server.js
